test: 사업자등록증 업로드 횟수 테스트 수정
