+++ src/test/java/com/birca/bircabackend/command/cafe/domain/OcrRequestHistoryTest.java
