+++ src/test/java/com/birca/bircabackend/support/enviroment/DocumentationTest.java
@@ -8,7 +8,7 @@ import com.birca.bircabackend.command.cafe.application.BusinessLicenseService;
 import com.birca.bircabackend.command.auth.application.AuthFacade;
 import com.birca.bircabackend.command.auth.application.token.JwtTokenProvider;
 import com.birca.bircabackend.command.cafe.domain.OcrRequestHistoryRepository;
-import com.birca.bircabackend.command.cafe.validation.OcrRequestCountValidator;
+import com.birca.bircabackend.command.cafe.infrastructure.uploadcount.OcrRequestCountValidator;
 import com.birca.bircabackend.command.member.application.MemberService;
 import com.birca.bircabackend.common.exception.ErrorCode;
 import com.birca.bircabackend.query.service.ArtistGroupQueryService;
