+++ src/test/resources/fixture/ocr-request-history-fixture.sql
@@ -1,0 +1,2 @@
+INSERT INTO ocr_request_history (owner_id, upload_count)
+VALUES (1, 5);
